requirements.txt
aiofiles==23.2.1
aiohttp==3.8.5
altair==5.0.1
annotated-types==0.5.0
anyio==3.7.1
asgiref==3.8.1
bcrypt==4.1.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
fastapi==0.109.0
ffmpy==0.3.1
groq==0.4.1
httpcore==1.0.1
httpx==0.26.0
idna==3.4
langchain>=0.1.0
langchain-groq>=0.1.0
langsmith>=0.0.77
matplotlib==3.8.2
motor==3.3.2
numpy==1.26.3
openai==0.27.9
pdf2image==1.17.0
Pillow==9.5.0
pydantic==2.6.0
PyPDF2==3.0.1
pymongo==4.6.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytesseract==0.3.10
reportlab==4.0.8
requests==2.31.0
starlette==0.27.0
streamlit==1.31.1
typing_extensions==4.8.0
uvicorn==0.27.0
watchfiles==0.21.0
